pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "athenaeum-kb"
version = "0.1.0"
description = "Tools for intelligent interaction with knowledge bases"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "langchain-core>=0.3",
    "langchain-chroma>=0.2",
    "langchain-openai>=0.3",
    "rank-bm25>=0.2.2",
    "markitdown>=0.1",
]

[project.optional-dependencies]
docling = ["docling>=2.0"]
mistral = ["mistralai>=1.0"]
lighton = ["transformers>=4.40", "torch>=2.0", "Pillow>=10.0"]
all-ocr = [
    "athenaeum-kb[docling]",
    "athenaeum-kb[mistral]",
    "athenaeum-kb[lighton]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/athenaeum"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "rank_bm25",
    "rank_bm25.*",
    "mistralai",
    "mistralai.*",
    "transformers",
    "transformers.*",
    "PIL",
    "PIL.*",
    "docling",
    "docling.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
